requirements-dev.txt
attrs==17.4.0
click==6.7
coverage==4.5.1
flask-graphql==1.4.1
flask==0.12.2
graphql-core==2.0
itsdangerous==0.24
jinja2==2.10
markupsafe==1.0
mock==2.0.0
mongomock==3.9.0
more-itertools==4.1.0
pbr==4.0.2
pluggy==0.6.0
promise==2.1
py==1.5.3
pytest-cov==2.5.1
pytest==3.5.0
pytz==2018.4
rx==1.6.1
sentinels==1.0.0
six==1.11.0
typing==3.6.4
werkzeug==0.14.1